Implement the TR_StartPC relocation on power
The TR_StartPC relocation added in eclipse/omr#6956 is now fully
implemented on power, the only platform to which it is relevant. A
manually-created load of the startPC that is created during code gen
under the EDO optimization now has an associated TR_StartPC record
added so that it is properly relocated.