Enable EDO under AOT and JITServer
Full support has been added for relocating the data and code related to
EDO, so this optimization is safe to run during AOT compilations and
when using the JITServer. Occasional segfaults still occur on power
with EDO and JITServer, so this optimization is disabled on that
platform until the cause can be determined.

Signed-off-by: Christian Despres <despresc@ibm.com>