Use a base JVM frontend for JITServer relocation
A new function to set the internal frontend of a TR::Compilation object
is used to temporarily override that frontend during method relocation
at a JITServer client after receiving an AOT from the server. This
ensures that the relocation runtime and SVM do not treat the method as
if it were freshly compiled at the client.

Signed-off-by: Christian Despres <despresc@ibm.com>