Correct string class initialization SVM validation
When checking whether or not the java/lang/String class is initialized
in the getStringClassEnableCompressionFieldAddr frontend query, we no
longer request that a TR_ValidateClassInfoIsInitialized relocation
record be generated by the TR_PersistentClassInfo::isInitialized method
that is called. The getStringClassEnableCompressionFieldAddr query is
only ever called during relocation, and so the compilation should never
need to generate that relocation record at that time. Formerly, the
getStringClassEnableCompressionFieldAddr method would
always request that such validation be performed. This behaviour, in
conjunction with some implementation details of
TR_PersistentClassInfo::isInitialized, had the undesirable side effect
of preventing the immediate relocation of AOT methods compiled with SVM
received by a JITServer client from a JITServer AOT cache.

Signed-off-by: Christian Despres <despresc@ibm.com>