Modify criteria for JITServer AOT load
A JITServer client will now request an AOT load even if a local SCC is
not present. When the relocation mechanism and deserializer have been
modified appropriately, this will allow JITServer AOT cache loads to
occur in that circumstance. Both the server and client will now
recognize AOT cache load requests even if the method is not eligible for
an AOT cache store.

Signed-off-by: Christian Despres <despresc@ibm.com>