Add option for new JITServer AOT cache
The -XX:[+|-]JITServerAOTCacheIgnoreLocalSCC option explicitly
enables or disables the use of the local SCC when compiling a method
using the JITServer AOT cache. If disabled, a local SCC must be present
and have sufficient writable space for the JITServer AOT cache to be
used. By default this option is enabled.

Signed-off-by: Christian Despres <despresc@ibm.com>