Add compilation property for client ignoring SCC
The _ignoringLocalSCC property is true at a JITServer client when the
compilation is to be stored in the JITServer's AOT cache and the client
is using the new deserializer implementation (hence ignoring its local
SCC). Since the server is responsible for generating relocation and
AOT cache records in these circumstances, this property is used to stop
certain SCC-related checks from being performed and stop relo records
from being generated when it is true.

Signed-off-by: Christian Despres <despresc@ibm.com>