Add initial server connection flag
The new _hasConnectedToServer flag in the persistent info is true at a
JITServer client when it has previously connected to a server,
regardless of the current status of any server connection. Tracking
this property is necessary to ensure that the deserializer is reset
on new connections, as the previous method (checking if the serverUID is
0) is not reliable for that purpose.

This fixes a bug that prevented the deserializer from being reset when
a client connected to a new server.

Signed-off-by: Christian Despres <despresc@ibm.com>