Improve port generation heuristic in JITServer test
We now request an unused port directly in the Java code, and give this
port to the client and server for use during testing. This is still not
a totally reliable way of getting a port for testing, but is at least an
improvement on the (not particularly accurate) method that was used
before.

Signed-off-by: Christian Despres <despresc@ibm.com>