Check loader validity during deserialization
Even if a class loader serialization record has already been
deserialized and cached in the JITServer AOT deserializer, we must still
check that the entry has not been invalidated by the unloading of the
associated class loader. This is required by the
TR_J9DeserializerSharedCache, as it must be able to return a cached
class loader in pointerFromOffsetInSharedCache if that function is given
an offset to a class loader record.

Note that this check only needs to be done when the client is ignoring
the local shared class cache - if the client is not ignoring its SCC,
then during relocation it will use offsets into the local SCC that
cannot become invalid.

Signed-off-by: Christian Despres <despresc@ibm.com>