Fix JITServer AOT deserializer revalidation
When deserializing a method from the JITServer AOT cache, the client, if
ignoring its local SCC, must ensure that all serialization records
referred to by that method are still valid. This is necessary to fulfill
the requests for cached data that the TR_J9DeserializerSharedCache will
make of the deserializer during the relocation of that method.

Signed-off-by: Christian Despres <despresc@ibm.com>