Fix JITServer AOT deserializer method invalidation
In the class redefinition event handler, the "old class" may already
have new RAM methods installed in it if the redefinition is happening
in-place; if so, the old methods will be in the "new class". To make
sure we invalidate all the cached methods of the class being redefined,
we now invalidate both the old and new class.

Signed-off-by: Christian Despres <despresc@ibm.com>