Allocate vector, mask names at JITServer startup
The vector and mask name arrays (used in the generation of logs) are
initialized exactly once. Formerly, this would always be done on-demand,
during a compilation. This caused memory corruption errors and
segfaulting at the server, as persistent allocations in a compilation
using the global allocator will, by default, use per-client memory. That
per-client memory is marked free when a client disconnects, so the name
array entries would become corrupted or inaccessible as a result.

Instead, the vector and mask name arrays are initialized at startup when
in server mode.

Signed-off-by: Christian Despres <despresc@ibm.com>