Track loaded classes with valid class chains
The new TR_AOTDependencyTable tracks all loaded classes whose class
chains match what is stored in the local SCC. In future, it will also be
responsible for tracking potential AOT loads; it will use this offset
map to detect when all the dependencies of the stored AOT compilation
have been satisfied and reduce the count of the associated method to
induce a load in response.

This class tracking is disabled by default. It can be controlled with
the option -XX:[+|-]TrackAOTDependencies.

Signed-off-by: Christian Despres <despresc@ibm.com>