Track methods with AOT bodies with dependencies
When a method is initialized in jitHookInitializeSendTarget, the SCC
will be checked to see if it has a stored AOT body that was compiled
with tracked dependencies. If it does, and all of them are satisfied,
the method will be given an initial count of 0. If not all of the
dependencies are satisfied, the method will be tracked by the
TR_DependencyTable until the dependencies are satisfied, at which point
its count will be set to zero, triggering an AOT load on its next
invocation.

Signed-off-by: Christian Despres <despresc@ibm.com>